requirements.txt
selenium
requests
python-dotenv
pillow